[#27] EditTimeViewModel 추가
