+++ TIG/Sources/App/AppViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  AppViewModel.swift
+//  TIG
+//
+//  Created by 신승재 on 4/15/25.
+//
+
+import Foundation
