+++ TIG/Sources/Feature/Home/ViewModel/HomeViewModel.swift
@@ -23,9 +23,7 @@ final class HomeViewModel {
     
     // 탭바 액션
     case changeTab(HomeTab)
-    
-    // 시간 수정 액션
-    case dailyTimeSaveTapped([TimeSlot])
+
   }
   
   private(set) var sharedState: SharedState = DIContainer.shared.resolve()
@@ -48,9 +46,6 @@ final class HomeViewModel {
       
     case .changeTab(let tab):
       state.selectedTab = tab
-      
-    case .dailyTimeSaveTapped(let timeSlots):
-      updateTimeSlot(date: sharedState.selectedDate, timeSlots: timeSlots)
     }
   }
 }
@@ -133,18 +128,5 @@ private extension HomeViewModel {
     } catch {
       return .failure(error)
     }
-  }
-  
-  /// TimeSlots을 업데이트 합니다.
-  /// - Parameters:
-  ///   - date: 업데이트할 날짜.
-  ///   - timeSlots: 해당 일정에 적용할 `TimeSlot` 배열.
-  func updateTimeSlot(date: Date, timeSlots: [TimeSlot]) {
-    dailyScheduleRepository.updateDailySchedule(
-      date: sharedState.selectedDate,
-      timeSlots: timeSlots
-    )
-    sharedState.groupedTimeSlots = timeSlots.groupedTimeSlots
-    sharedState.timeSlots = timeSlots
   }
 }
